# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Utilisateurs",
        "description": "Cette API permet de gérer les utilisateurs, leurs profils et leurs informations d'authentification.",
        "contact": {
            "email": "yvesedi41@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.252.148:8000/api",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
        "/create_emergency": {
            "post": {
                "tags": [
                    "Emergency"
                ],
                "summary": "Create a new emergency",
                "description": "Crée une nouvelle entrée pour une urgence et un utilisateur associé.",
                "operationId": "771bbd8f00f82ff09e16d371665d9eca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "emplacement",
                                    "email",
                                    "phone_number",
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Urgence Médicale"
                                    },
                                    "emplacement": {
                                        "type": "string",
                                        "example": "123 Rue Exemple, Paris"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "urgence@example.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0612345678"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "emergencyuser"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "securepassword"
                                    },
                                    "logo": {
                                        "description": "Image file for the emergency's logo",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Success"
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "field_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The field_name field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/actualise": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Actualiser les informations de l'utilisateur",
                "operationId": "fc72ceec556b052cb6a712fce643a92d",
                "responses": {
                    "200": {
                        "description": "Informations de l'utilisateur actualisées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Enregistrer un nouvel utilisateur",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+123456789"
                                    },
                                    "sex": {
                                        "type": "string",
                                        "example": "male"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "blood_type": {
                                        "type": "string",
                                        "example": "O+"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "example": "1990-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inscription réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registration successful"
                                        },
                                        "user": {
                                            "type": "object"
                                        },
                                        "access_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation échouée"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Connecter un utilisateur",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "user": {
                                            "type": "object"
                                        },
                                        "access_token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                }
            }
        },
        "/username_verify": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Vérifier la disponibilité du nom d'utilisateur",
                "operationId": "05b3ccfe892dc2aab1f8ddf94a40caef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "johndoe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nom d'utilisateur disponible"
                    },
                    "400": {
                        "description": "Nom d'utilisateur déjà pris"
                    }
                }
            }
        },
        "/register_member": {
            "post": {
                "tags": [
                    "Emergency"
                ],
                "summary": "Register a new member",
                "description": "Enregistre un nouvel utilisateur avec les informations fournies.",
                "operationId": "7ad9c677640b4130a44d53934386f961",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "matricule",
                                    "phone_number",
                                    "position",
                                    "email",
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "matricule": {
                                        "type": "string",
                                        "example": "M123456"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "0612345678"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "Developer"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "photo": {
                                        "description": "Image file for the user's photo",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur ajouté avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Utilisateur ajouté avec succès"
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "field_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The field_name field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Emergency",
            "description": "Emergency"
        },
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}